Fixed LCD initialization in initLCD fundtion. LCD manual insructs to turn display off, perform a few commands, then turn it on. Turning it off caused the initialization to fail -- unknown why.
Did this version work and then not work? There was a small commit checked
into Git after this version, implying that it worked at least once. ???

Turned off brownout reset -- code doesn't currently know how to handle
this so if a brownout occurs a total reset of the software is triggered.

Updated author's initials and version number displayed on startup.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>