Display on/off Cursor Blink command is no longer sent directly to the display when received from the "Main" PIC. The code is stored and applied at the end of a display update.
The last cursor location received from the "Main" PIC is now stored.

When the blink was on, it could be seen running across the screen during
an update as it followed each new character being written.

In this code, the character at the last selected cursor location does NOT
blink so the code is unusable. Need to make the character blink at the
end of a refresh cycle in some way.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>